+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Released on Sunday, July 26 2026.
 - Added missing shape declarations (e.g., `shape-outside`, `shape-rendering`)
 - Added missing anchor positioning declarations (e.g., `anchor-name`, `anchor-scope`, `position-anchor`)
 - Added missing text declarations (e.g., `text-underline-offset`, `tab-size`, `initial-letter`)
+- Added missing masking declarations (e.g., `mask-image`, `mask-border-source`)
 - Added missing auxiliary declarations (e.g., `counter-set`, `image-rendering`, `view-transition-name`)
 - Added more CSSOM possibilities and helpers (#6)
 - Added parts of recent color spec update such as `rgb` with spaces (#131)